+++ addOns/quickstart/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Add Scan Policy option to the Automated Scan panel.
 
 
 ## [55] - 2026-03-09